structural: Extract local segment storage core
