+++ streamarr-transcode-engine/src/main/java/com/streamarr/transcode/engine/segment/InvalidSegmentPathException.java
@@ -1,0 +1,8 @@
+package com.streamarr.transcode.engine.segment;
+
+public class InvalidSegmentPathException extends RuntimeException {
+
+  public InvalidSegmentPathException(String segment) {
+    super("Invalid segment path: " + segment);
+  }
+}
